--- src/client.html
@@ -1,13 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <title>Arca</title>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/2.0.4/socket.io.js"></script>
-    <link href="/static/app.bundle.css" rel="stylesheet" type="text/css"></link>
-  </head>
-  <body>
-    <!-- <script src="client.js"></script> -->
-    <div id="root"></div>
-    <script src="/static/bundle.js"></script>
-  </body>
-</html>
