--- src/store/configureStore.js
@@ -1,8 +1,0 @@
-import {createStore, applyMiddleware} from 'redux';
-import rootReducer from 'src/reducers/index';
-import thunk from 'redux-thunk';
-
-export default function configureStore(initialState) {
-  const store = createStore(rootReducer, initialState, applyMiddleware(thunk));
-  return store;
-}
