--- src/styles/app.scss
@@ -1,9 +1,0 @@
-.row {
-    margin-bottom: 15px;
-
-    span {
-        border: 1px solid red;
-        padding: 15px;
-        display: inline-block;
-    }
-}
