+++ static/client.html
@@ -1,0 +1,13 @@
+<!doctype html>
+<html>
+  <head>
+    <title>Arca</title>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/2.0.4/socket.io.js"></script>
+    <link href="/dist/app.bundle.css" rel="stylesheet" type="text/css"></link>
+  </head>
+  <body>
+    <!-- <script src="client.js"></script> -->
+    <div id="root"></div>
+    <script src="/dist/bundle.js"></script>
+  </body>
+</html>
