+++ static/store/configureStore.js
@@ -1,0 +1,8 @@
+import {createStore, applyMiddleware} from 'redux';
+import rootReducer from 'static/reducers/index';
+import thunk from 'redux-thunk';
+
+export default function configureStore(initialState) {
+  const store = createStore(rootReducer, initialState, applyMiddleware(thunk));
+  return store;
+}
