+++ static/styles/app.scss
@@ -1,0 +1,9 @@
+.row {
+    margin-bottom: 15px;
+
+    span {
+        border: 1px solid red;
+        padding: 15px;
+        display: inline-block;
+    }
+}
